--- fluid-cpp/fluid_nl.h
@@ -1,29 +1,0 @@
-#ifndef FLUID_NL_H_
-#define FLUID_NL_H_
-#include <vector>
-
-#define PI 3.14159265359
-
-int fluid_nl(double* crossSectionLength,
-             double* velocity,
-             double* pressure,
-             double t,
-             int N,
-             double kappa,
-             double tau);
-
-int linsolve(int n,
-             double** A,
-             double* b,
-             double* x);
-             
-void write_vtk(double t, 
-				int iteration, 
-				const char* filename_prefix,
-				int N_slices,
-				double* grid,
-				double* velocity, 
-				double* pressure, 
-				double* diameter);             
-
-#endif
